+++ src/main/java/net/frozenblock/lib/particle/ColoredCampfireSmokeParticle.java
@@ -1,0 +1,101 @@
+package net.frozenblock.lib.particle;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.frozenblock.lib.particle.options.ColoredSmokeParticleOptions;
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.particle.CampfireSmokeParticle;
+import net.minecraft.client.particle.Particle;
+import net.minecraft.client.particle.ParticleProvider;
+import net.minecraft.client.particle.SpriteSet;
+import net.minecraft.client.renderer.texture.TextureAtlasSprite;
+import net.minecraft.util.Mth;
+import net.minecraft.util.RandomSource;
+
+@Environment(EnvType.CLIENT)
+public class ColoredCampfireSmokeParticle extends CampfireSmokeParticle {
+	private final float targetRColor;
+	private final float startRColor;
+	private final float targetBColor;
+	private final float startBColor;
+	private final float targetGColor;
+	private final float startGColor;
+	private final int colorLerpEndsAt;
+	private int colorLerpTicks;
+
+	protected ColoredCampfireSmokeParticle(
+		ClientLevel level,
+		double x, double y, double z,
+		double xa, double ya, double za,
+		float rDifference, float gDifference, float bDifference,
+		boolean isSignalFire,
+		TextureAtlasSprite sprite
+	) {
+		super(level, x, y, z, xa, ya, za, isSignalFire, sprite);
+		this.targetRColor = this.rCol;
+		this.rCol = this.startRColor = Math.clamp(this.targetRColor + rDifference, 0F, 1F);
+		this.targetGColor = this.gCol;
+		this.gCol = this.startGColor = Math.clamp(this.targetGColor + gDifference, 0F, 1F);
+		this.targetBColor = this.bCol;
+		this.bCol = this.startBColor = Math.clamp(this.targetBColor + bDifference, 0F, 1F);
+		this.colorLerpEndsAt = level.getRandom().nextInt(20, 40);
+	}
+
+	@Override
+	public void tick() {
+		this.colorLerpTicks += 1;
+		super.tick();
+	}
+
+	@Override
+	protected int getLightCoords(float partialTick) {
+		final float colorLerp = Math.min((this.colorLerpTicks + partialTick), this.colorLerpEndsAt) / this.colorLerpEndsAt;
+		this.rCol = Mth.lerp(colorLerp, this.startRColor, this.targetRColor);
+		this.gCol = Mth.lerp(colorLerp, this.startGColor, this.targetGColor);
+		this.bCol = Mth.lerp(colorLerp, this.startBColor, this.targetBColor);
+		return super.getLightCoords(partialTick);
+	}
+
+	public record CosyProvider(SpriteSet spriteSet) implements ParticleProvider<ColoredSmokeParticleOptions> {
+		@Override
+		public Particle createParticle(
+			ColoredSmokeParticleOptions options,
+			ClientLevel level,
+			double x, double y, double z,
+			double xAux, double yAux, double zAux,
+			RandomSource random
+		) {
+			final ColoredCampfireSmokeParticle particle = new ColoredCampfireSmokeParticle(
+				level,
+				x, y, z,
+				xAux, yAux, zAux,
+				options.rDifference(), options.gDifference(), options.bDifference(),
+				false,
+				this.spriteSet.get(random)
+			);
+			particle.setAlpha(0.9F);
+			return particle;
+		}
+	}
+
+	public record SignalProvider(SpriteSet spriteSet) implements ParticleProvider<ColoredSmokeParticleOptions> {
+		@Override
+		public Particle createParticle(
+			ColoredSmokeParticleOptions options,
+			ClientLevel level,
+			double x, double y, double z,
+			double xAux, double yAux, double zAux,
+			RandomSource random
+		) {
+			final ColoredCampfireSmokeParticle particle = new ColoredCampfireSmokeParticle(
+				level,
+				x, y, z,
+				xAux, yAux, zAux,
+				options.rDifference(), options.gDifference(), options.bDifference(),
+				true,
+				this.spriteSet.get(random));
+			particle.setAlpha(0.95F);
+			return particle;
+		}
+	}
+}